requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
email-validator==2.1.0
python-dotenv==1.0.0
google-generativeai==0.3.0
pyjwt==2.9.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0
pydantic-extra-types==2.4.1
pymongo[srv]==4.6.1